Release of 0.2.0 -- The first full multi-platform working version of umpire

+++ umpire/unpack.py
@@ -1,7 +1,7 @@
 """unpack.py"""
 
 HELPTEXT = """
-                  ----- Unpack Module -----
+     ----- Unpack Module -----
 
 The unpack module unpacks a tar gz file.
 
@@ -14,6 +14,9 @@ from maestro.core import module
 
 HELP_KEYS = ["h", "help"]
 
+class UnpackError(Exception):
+    pass
+
 class UnpackModule(module.AsyncModule):
     # Required ID of this module
     id = "unpack"
@@ -22,13 +25,15 @@ class UnpackModule(module.AsyncModule):
     delete_archive = False
 
     def help(self):
-        print self.help_text
+        print(self.help_text)
         exit(0)
 
     def run(self,kwargs):
-        import tarfile
-        with tarfile.open(self.file_path, "r:gz") as f:
-            f.extractall(self.destination_path)
-        if self.delete_archive is True:
-            os.remove(self.file_path)
-
+        if self.file_path.endswith("tar.gz") or self.file_path.endswith("tgz"):
+            import tarfile
+            with tarfile.open(self.file_path, "r:gz") as f:
+                f.extractall(self.destination_path)
+            if self.delete_archive is True:
+                os.remove(self.file_path)
+        else:
+            raise UnpackError("Unable to unpack this type of file: " + os.path.split(self.file_path)[1])
